etc(#153): dev 결제 로직 병합
