+++ src/app/class/[id]/apply/_components/Coupon/CouponClient.tsx
